+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Breaking
 
 Changed
   * AZ-890: Use SSD disks by default
+  * AZ-890: Use ephemeral storage for OS on temp disk by default 
   * AZ-890: Change some variables default value
 
 # v6.3.0 - 2022-11-04